+++ src-cpp/master/master.h
@@ -5,7 +5,7 @@
 #include <iostream>
 #include <fstream>
 #include "asio.hpp"
-#include "util/mpacket.h"
+#include "net/mpacket.h"
 #include "net/time_packet.h"
 #include <sndfile.hh>
