+++ src-cpp/net/packet.cc
@@ -1,11 +1,12 @@
 #include "net/packet.h"
 #include "net/time_packet.h"
-#include "util/mpacket.h"
+#include "net/mpacket.h"
 
 Packet::Packet() {
 }
 
-Packet::Packet(PacketType pt) : packet_type(pt) { 
+Packet::Packet(PacketType pt) : packet_type(pt) {
+
 }
 
 PacketType Packet::get_type(){
@@ -21,7 +22,7 @@ Packet* Packet::unpack(uint8_t* buf, std::size_t size){
         return MPacket::unpack(buf + 1,size - 1);
       default:
         std::cerr << "Data corruption detected: Invalid flag." << std::endl;
+        return nullptr;
     }
 }
-
